MTDSA-31536: Add new 25/26 schema for API#1952 Create and Amend CGT Residential Property Disposals (non-PPD)
