+++ conf/v3.routes
@@ -2,7 +2,7 @@
 PUT        /residential-property/:nino/:taxYear/ppd               v3.residentialPropertyDisposals.createAmendCgtPpdOverrides.CreateAmendCgtPpdOverridesController.createAmendCgtPpdOverrides(nino: String, taxYear: String)
 DELETE     /residential-property/:nino/:taxYear/ppd               v3.residentialPropertyDisposals.deleteCgtPpdOverrides.DeleteCgtPpdOverridesController.deleteCgtPpdOverrides(nino: String, taxYear: String)
 
-PUT        /disposals-income/residential-property/:nino/:taxYear  v3.residentialPropertyDisposals.createAmendNonPpd.CreateAmendCgtResidentialPropertyDisposalsController.createAmendCgtResidentialPropertyDisposals(nino: String, taxYear: String)
+PUT        /residential-property/:nino/:taxYear                   v3.residentialPropertyDisposals.createAmendNonPpd.CreateAmendCgtResidentialPropertyDisposalsController.createAmendCgtResidentialPropertyDisposals(nino: String, taxYear: String)
 GET        /residential-property/:nino/:taxYear                   v3.residentialPropertyDisposals.retrieveNonPpd.RetrieveCgtResidentialPropertyController.retrieve(nino: String, taxYear: String)
 
 PUT        /other-gains/:nino/:taxYear                            v3.otherCgt.createAmend.CreateAmendOtherCgtController.createAmendOtherCgt(nino: String, taxYear: String)
