--- it/test/v3/endpoints/Def2_CreateAmendCgtResidentialPropertyDisposalsControllerHipISpec.scala
@@ -1,489 +1,0 @@
-/*
- * Copyright 2025 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package v3.endpoints
-
-import com.github.tomakehurst.wiremock.client.WireMock.*
-import com.github.tomakehurst.wiremock.stubbing.StubMapping
-import common.errors.*
-import play.api.http.HeaderNames.ACCEPT
-import play.api.http.Status.*
-import play.api.libs.json.{JsObject, JsValue, Json}
-import play.api.libs.ws.WSBodyWritables.writeableOf_JsValue
-import play.api.libs.ws.{WSRequest, WSResponse}
-import play.api.test.Helpers.AUTHORIZATION
-import shared.models.errors.*
-import shared.services.*
-import shared.support.{IntegrationBaseSpec, WireMockMethods}
-
-class Def2_CreateAmendCgtResidentialPropertyDisposalsControllerHipISpec extends IntegrationBaseSpec with WireMockMethods {
-
-  val validDisposalDate: String    = "2020-03-27"
-  val validCompletionDate: String  = "2020-03-29"
-  val validAcquisitionDate: String = "2020-03-25"
-
-  val validRequestJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "customerReference": "CGTDISPOSAL01",
-      |         "disposalDate": "$validDisposalDate",
-      |         "completionDate": "$validCompletionDate",
-      |         "disposalProceeds": 1999.99,
-      |         "acquisitionDate": "$validAcquisitionDate",
-      |         "acquisitionAmount": 1999.99,
-      |         "improvementCosts": 1999.99,
-      |         "additionalCosts": 1999.99,
-      |         "prfAmount": 1999.99,
-      |         "otherReliefAmount": 1999.99,
-      |         "lossesFromThisYear": 1999.99,
-      |         "lossesFromPreviousYear": 1999.99,
-      |         "amountOfNetGain": 1999.99,
-      |         "numberOfDisposals": 3,
-      |         "gainsWithBadr": 99999999999.99,
-      |         "gainsBeforeLosses": 99999999999.99,
-      |         "claimOrElectionCodes": ["PRR"]
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val noMeaningfulDataJson: JsValue = Json.parse(
-    """
-      |{
-      |  "aField": "aValue"
-      |}
-       """.stripMargin
-  )
-
-  val emptyDisposalsJson: JsValue = Json.parse(
-    """
-      |{
-      |   "disposals": []
-      |}
-     """.stripMargin
-  )
-
-  val emptyDisposalsError: MtdError = RuleIncorrectOrEmptyBodyError.copy(
-    paths = Some(Seq("/disposals"))
-  )
-
-  val missingFieldsJson: JsValue = Json.parse(
-    """
-      |{
-      |   "disposals": [{}]
-      |}
-     """.stripMargin
-  )
-
-  val missingFieldsError: MtdError = RuleIncorrectOrEmptyBodyError.copy(
-    paths = Some(
-      Seq(
-        "/disposals/0/acquisitionAmount",
-        "/disposals/0/acquisitionDate",
-        "/disposals/0/completionDate",
-        "/disposals/0/disposalDate",
-        "/disposals/0/disposalProceeds",
-        "/disposals/0/gainsBeforeLosses",
-        "/disposals/0/numberOfDisposals"
-      )
-    )
-  )
-
-  val decimalsTooBigJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "numberOfDisposals": 3,
-      |         "customerReference": "CGTDISPOSAL01",
-      |         "disposalDate": "$validDisposalDate",
-      |         "completionDate": "$validCompletionDate",
-      |         "disposalProceeds": 100000000000.00,
-      |         "acquisitionDate": "$validAcquisitionDate",
-      |         "acquisitionAmount": 100000000000.00,
-      |         "improvementCosts": 100000000000.00,
-      |         "additionalCosts": 100000000000.00,
-      |         "prfAmount": 100000000000.00,
-      |         "otherReliefAmount": 100000000000.00,
-      |         "lossesFromThisYear": 100000000000.00,
-      |         "lossesFromPreviousYear": 100000000000.00,
-      |         "gainsWithBadr": 100000000000.00,
-      |         "gainsBeforeLosses": 100000000000.00,
-      |         "amountOfNetGain": 100000000000.00
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val decimalsTooSmallJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "numberOfDisposals": 3,
-      |         "customerReference": "CGTDISPOSAL01",
-      |         "disposalDate": "$validDisposalDate",
-      |         "completionDate": "$validCompletionDate",
-      |         "disposalProceeds": -0.1,
-      |         "acquisitionDate": "$validAcquisitionDate",
-      |         "acquisitionAmount": -0.1,
-      |         "improvementCosts": -0.1,
-      |         "additionalCosts": -0.1,
-      |         "prfAmount": -0.1,
-      |         "otherReliefAmount": -0.1,
-      |         "lossesFromThisYear": -0.1,
-      |         "lossesFromPreviousYear": -0.1,
-      |         "gainsWithBadr": -0.1,
-      |         "gainsBeforeLosses": -0.1,
-      |         "amountOfNetGain": -0.1
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val DecimalsOutOfRangeError: MtdError = ValueFormatError.copy(
-    message = "The value must be between 0 and 99999999999.99",
-    paths = Some(
-      Seq(
-        "/disposals/0/disposalProceeds",
-        "/disposals/0/acquisitionAmount",
-        "/disposals/0/gainsBeforeLosses",
-        "/disposals/0/improvementCosts",
-        "/disposals/0/additionalCosts",
-        "/disposals/0/prfAmount",
-        "/disposals/0/otherReliefAmount",
-        "/disposals/0/lossesFromThisYear",
-        "/disposals/0/lossesFromPreviousYear",
-        "/disposals/0/amountOfNetGain",
-        "/disposals/0/gainsWithBadr"
-      ))
-  )
-
-  val datesNotFormattedJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "numberOfDisposals": 3,
-      |         "customerReference": "CGTDISPOSAL01",
-      |         "disposalDate": "202005-07",
-      |         "completionDate": "21-05-07",
-      |         "disposalProceeds": 1999.99,
-      |         "acquisitionDate": "2020-5-7",
-      |         "acquisitionAmount": 1999.99,
-      |         "improvementCosts": 1999.99,
-      |         "additionalCosts": 1999.99,
-      |         "prfAmount": 1999.99,
-      |         "otherReliefAmount": 1999.99,
-      |         "lossesFromThisYear": 1999.99,
-      |         "lossesFromPreviousYear": 1999.99,
-      |         "gainsWithBadr": 1999.99,
-      |         "gainsBeforeLosses": 1999.99,
-      |         "amountOfNetGain": 1999.99
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val datesNotFormattedError: MtdError = DateFormatError.copy(
-    message = "The supplied date format is not valid",
-    paths = Some(
-      Seq(
-        "/disposals/0/disposalDate",
-        "/disposals/0/completionDate",
-        "/disposals/0/acquisitionDate"
-      ))
-  )
-
-  val customerRefTooLongJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "numberOfDisposals": 3,
-      |         "customerReference": "ThisIs91CharactersLongCGTDIS000000000000000000000000000000000000000000000000000000000000001",
-      |         "disposalDate": "$validDisposalDate",
-      |         "completionDate": "$validCompletionDate",
-      |         "disposalProceeds": 1999.99,
-      |         "acquisitionDate": "$validAcquisitionDate",
-      |         "acquisitionAmount": 1999.99,
-      |         "improvementCosts": 1999.99,
-      |         "additionalCosts": 1999.99,
-      |         "prfAmount": 1999.99,
-      |         "otherReliefAmount": 1999.99,
-      |         "lossesFromThisYear": 1999.99,
-      |         "lossesFromPreviousYear": 1999.99,
-      |         "gainsWithBadr": 1999.99,
-      |         "gainsBeforeLosses": 1999.99,
-      |         "amountOfNetGain": 1999.99
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val customerRefTooShortJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "numberOfDisposals": 3,
-      |         "customerReference": "",
-      |         "disposalDate": "$validDisposalDate",
-      |         "completionDate": "$validCompletionDate",
-      |         "disposalProceeds": 1999.99,
-      |         "acquisitionDate": "$validAcquisitionDate",
-      |         "acquisitionAmount": 1999.99,
-      |         "improvementCosts": 1999.99,
-      |         "additionalCosts": 1999.99,
-      |         "prfAmount": 1999.99,
-      |         "otherReliefAmount": 1999.99,
-      |         "lossesFromThisYear": 1999.99,
-      |         "lossesFromPreviousYear": 1999.99,
-      |         "gainsWithBadr": 1999.99,
-      |         "gainsBeforeLosses": 1999.99,
-      |         "amountOfNetGain": 1999.99
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val customerRefError: MtdError = CustomerRefFormatError.withPath("/disposals/0/customerReference")
-
-  val gainLossJson: JsValue = Json.parse(
-    s"""
-      |{
-      |   "disposals":[
-      |      {
-      |         "numberOfDisposals": 3,
-      |         "customerReference": "CGTDISPOSAL01",
-      |         "disposalDate": "$validDisposalDate",
-      |         "completionDate": "$validCompletionDate",
-      |         "disposalProceeds": 1999.99,
-      |         "acquisitionDate": "$validAcquisitionDate",
-      |         "acquisitionAmount": 1999.99,
-      |         "improvementCosts": 1999.99,
-      |         "additionalCosts": 1999.99,
-      |         "prfAmount": 1999.99,
-      |         "otherReliefAmount": 1999.99,
-      |         "lossesFromThisYear": 1999.99,
-      |         "lossesFromPreviousYear": 1999.99,
-      |         "amountOfNetGain": 19999.99,
-      |         "gainsWithBadr": 1999.99,
-      |         "gainsBeforeLosses": 1999.99,
-      |         "amountOfNetLoss": 19999.99
-      |      }
-      |   ]
-      |}
-     """.stripMargin
-  )
-
-  val numberOfDisposalsJson: JsValue = Json.parse(
-    s"""
-       |{
-       |   "disposals":[
-       |      {
-       |         "numberOfDisposals": 0,
-       |         "customerReference": "CGTDISPOSAL01",
-       |         "disposalDate": "$validDisposalDate",
-       |         "completionDate": "$validCompletionDate",
-       |         "disposalProceeds": 1999.99,
-       |         "acquisitionDate": "$validAcquisitionDate",
-       |         "acquisitionAmount": 1999.99,
-       |         "improvementCosts": 1999.99,
-       |         "additionalCosts": 1999.99,
-       |         "prfAmount": 1999.99,
-       |         "otherReliefAmount": 1999.99,
-       |         "lossesFromThisYear": 1999.99,
-       |         "lossesFromPreviousYear": 1999.99,
-       |         "gainsBeforeLosses": 123.43,
-       |         "amountOfNetGain": 1999.99
-       |      }
-       |   ]
-       |}
-     """.stripMargin
-  )
-
-  val gainLossError: MtdError = RuleGainLossError.copy(
-    message = "Only one of gain or loss values can be provided",
-    paths = Some(
-      Seq(
-        "/disposals/0"
-      ))
-  )
-
-  val numberOfDisposalsError: MtdError = RuleIncorrectLossesSubmittedError.copy(
-    message = "LossesFromThisYear can not be submitted if the numberOfDisposals is less than or equal to 1",
-    paths = Some(
-      Seq(
-        "/disposals/0"
-      ))
-  )
-
-  trait Test {
-
-    val nino: String    = "AA123456A"
-    def taxYear: String = "2025-26"
-
-    def downstreamUri: String = s"/itsa/income-tax/v1/25-26/income/disposals/residential-property/$nino"
-
-    def setupStubs(): StubMapping
-
-    def request: WSRequest = {
-      setupStubs()
-      buildRequest(s"/disposals-income/residential-property/$nino/$taxYear")
-        .withHttpHeaders(
-          (ACCEPT, "application/vnd.hmrc.3.0+json"),
-          (AUTHORIZATION, "Bearer 123") // some bearer token
-        )
-    }
-
-    def verifyNrs(payload: JsValue): Unit =
-      verify(
-        postRequestedFor(urlEqualTo(s"/mtd-api-nrs-proxy/$nino/itsa-cgt-disposal"))
-          .withRequestBody(equalToJson(payload.toString())))
-
-  }
-
-  "Calling the 'create and amend other CGT' endpoint" should {
-    "return a 200 status code" when {
-      "any valid request is made for a TYS tax year" in new Test {
-
-        override def setupStubs(): StubMapping = {
-          AuditStub.audit()
-          AuthStub.authorised()
-          MtdIdLookupStub.ninoFound(nino)
-          DownstreamStub.onSuccess(DownstreamStub.PUT, downstreamUri, NO_CONTENT)
-        }
-
-        val response: WSResponse = await(request.put(validRequestJson))
-        response.status shouldBe OK
-
-        verifyNrs(validRequestJson)
-      }
-    }
-
-    "return error according to spec" when {
-      "validation error" when {
-        def validationErrorTest(requestNino: String,
-                                requestTaxYear: String,
-                                requestBody: JsValue,
-                                expectedStatus: Int,
-                                expectedError: MtdError,
-                                expectedErrors: Option[ErrorWrapper],
-                                scenario: Option[String]): Unit = {
-          s"validation fails with ${expectedError.code} error${scenario.fold("")(scenario => s" for $scenario scenario")}" in new Test {
-
-            override val nino: String    = requestNino
-            override val taxYear: String = requestTaxYear
-
-            override def setupStubs(): StubMapping = {
-              AuditStub.audit()
-              AuthStub.authorised()
-              MtdIdLookupStub.ninoFound(nino)
-            }
-
-            val response: WSResponse = await(request.put(requestBody))
-            response.status shouldBe expectedStatus
-            response.json shouldBe expectedErrors.fold(Json.toJson(expectedError))(errorWrapper => Json.toJson(errorWrapper))
-            response.header("Content-Type") shouldBe Some("application/json")
-          }
-        }
-
-        val input = Seq(
-          // Path errors
-          ("AA1123A", "2025-26", validRequestJson, BAD_REQUEST, NinoFormatError, None, None),
-          ("AA123456A", "20177", validRequestJson, BAD_REQUEST, TaxYearFormatError, None, None),
-          ("AA123456A", "2015-17", validRequestJson, BAD_REQUEST, RuleTaxYearRangeInvalidError, None, None),
-          ("AA123456A", "2018-19", validRequestJson, BAD_REQUEST, RuleTaxYearNotSupportedError, None, None),
-
-          // Body errors
-          ("AA123456A", "2025-26", JsObject.empty, BAD_REQUEST, RuleIncorrectOrEmptyBodyError, None, Some("emptyBody")),
-          ("AA123456A", "2025-26", emptyDisposalsJson, BAD_REQUEST, emptyDisposalsError, None, Some("empty disposals")),
-          ("AA123456A", "2025-26", missingFieldsJson, BAD_REQUEST, missingFieldsError, None, Some("missing all mandatory fields")),
-          ("AA123456A", "2025-26", decimalsTooBigJson, BAD_REQUEST, DecimalsOutOfRangeError, None, Some("Decimals too big")),
-          ("AA123456A", "2025-26", decimalsTooSmallJson, BAD_REQUEST, DecimalsOutOfRangeError, None, Some("Decimals too small")),
-          ("AA123456A", "2025-26", datesNotFormattedJson, BAD_REQUEST, datesNotFormattedError, None, Some("incorrect date formats")),
-          ("AA123456A", "2025-26", customerRefTooLongJson, BAD_REQUEST, customerRefError, None, Some("bad customer reference")),
-          ("AA123456A", "2025-26", customerRefTooShortJson, BAD_REQUEST, customerRefError, None, Some("empty customer reference string")),
-          ("AA123456A", "2025-26", gainLossJson, BAD_REQUEST, gainLossError, None, Some("gain and loss provided")),
-          ("AA123456A", "2025-26", numberOfDisposalsJson, BAD_REQUEST, numberOfDisposalsError, None, Some("numberOfDisposals is less than 1"))
-        )
-        input.foreach(args => validationErrorTest.tupled(args))
-      }
-
-      "service error" when {
-        def serviceErrorTest(downstreamStatus: Int, downstreamCode: String, expectedStatus: Int, expectedBody: MtdError): Unit = {
-          s"downstream returns an $downstreamCode error and status $downstreamStatus" in new Test {
-
-            override def setupStubs(): StubMapping = {
-              AuditStub.audit()
-              AuthStub.authorised()
-              MtdIdLookupStub.ninoFound(nino)
-              DownstreamStub.onError(DownstreamStub.PUT, downstreamUri, downstreamStatus, errorBody(downstreamCode))
-            }
-
-            val response: WSResponse = await(request.put(validRequestJson))
-            response.status shouldBe expectedStatus
-            response.json shouldBe Json.toJson(expectedBody)
-            response.header("Content-Type") shouldBe Some("application/json")
-
-            verifyNrs(validRequestJson)
-          }
-        }
-
-        def errorBody(`type`: String): String =
-          s"""
-             |{
-             |  "origin": "HIP",
-             |  "response": {
-             |    "failures": [
-             |      {
-             |        "type": "${`type`}",
-             |        "reason": "downstream message"
-             |      }
-             |    ]
-             |  }
-             |}
-             |""".stripMargin
-
-        val errors = Seq(
-          (BAD_REQUEST, "INVALID_TAXABLE_ENTITY_ID", BAD_REQUEST, NinoFormatError),
-          (BAD_REQUEST, "INVALID_TAX_YEAR", BAD_REQUEST, TaxYearFormatError),
-          (BAD_REQUEST, "INVALID_PAYLOAD", INTERNAL_SERVER_ERROR, InternalError),
-          (UNPROCESSABLE_ENTITY, "INVALID_DISPOSAL_DATE", BAD_REQUEST, RuleDisposalDateErrorV1),
-          (UNPROCESSABLE_ENTITY, "INVALID_COMPLETION_DATE", BAD_REQUEST, RuleCompletionDateError),
-          (UNPROCESSABLE_ENTITY, "INVALID_ACQUISITION_DATE", BAD_REQUEST, RuleAcquisitionDateAfterDisposalDateError),
-          (UNPROCESSABLE_ENTITY, "OUTSIDE_AMENDMENT_WINDOW", BAD_REQUEST, RuleOutsideAmendmentWindowError),
-          (INTERNAL_SERVER_ERROR, "SERVER_ERROR", INTERNAL_SERVER_ERROR, InternalError),
-          (SERVICE_UNAVAILABLE, "SERVICE_UNAVAILABLE", INTERNAL_SERVER_ERROR, InternalError),
-          (UNPROCESSABLE_ENTITY, "TAX_YEAR_NOT_SUPPORTED", BAD_REQUEST, RuleTaxYearNotSupportedError),
-          (BAD_REQUEST, "INVALID_CORRELATION_ID", INTERNAL_SERVER_ERROR, InternalError)
-        )
-
-        errors.foreach(args => serviceErrorTest.tupled(args))
-      }
-    }
-  }
-
-}
